requirements.txt
azure-ai-textanalytics==5.2.1
azure-common==1.1.28
azure-core==1.26.4
azure-data-tables==12.4.2
azure-functions==1.14.0
azure-storage-blob==12.16.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
filelock==3.12.0
idna==3.4
isodate==0.6.1
lxml==4.9.2
msrest==0.7.1
multidict==6.0.4
oauthlib==3.2.2
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@786815dd05681e2421cd03aa9acf5ab5c773bce9
soupsieve==2.4.1
typing_extensions==4.6.2
urllib3==2.0.2
yarl==1.9.2
